fix: removed unnecessary files
Since we aren't using CMake (and if we wanted to we would just use the original repository and not this fork), it made sense to remove all of the CMake files.

Since we don't really care about examples or tests right now, it made sense to remove them as well. Since most of the "/deps/" directory seemed to be used for the examples and tests, the entire "/deps/" directory was removed as well. This has the side-effect of removing possible future support for Wayland platforms, but that can be a problem for later (if at all).

Lastly, documentation and some extra "build script/action" type files were removed since they were no longer needed.